Cmake file should support building both the cpp source as an executable and  the python extension module.
Building cpp source as an executable will be required to run the cpp tests.
Oython extension module must be discoverable, so that it can be imported in python. A separate folder python/ is created
to put the __init__.py file to make the module discoverable. We can use the name tudat. Putting tudat/ inside src/ will create
confusion, so we put it in a separate folder python/ in the root directory. This is the flder where tudatpy contents can go.